requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
mysql-connector-python==9.1.0
packaging==24.2
python-dotenv==1.0.1
